feat(machine-info): move nvidia info into machine-info component, rename to machine-info component
